move newfirmware/taus and newfirmware/common
