+++ L1Trigger/Phase2L1ParticleFlow/interface/taus/TauNNIdHW.h
@@ -7,9 +7,23 @@
 #include <complex>
 #include "ap_int.h"
 #include "ap_fixed.h"
-#include "parameters.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/tau_parameters.h"
 
 #include "DataFormats/L1TParticleFlow/interface/PFCandidate.h"
+
+#include "L1Trigger/Phase2L1ParticleFlow/interface/common/nnet_layer.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/common/nnet_activation.h"
+
+//hls-fpga-machine-learning insert weights
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/w1.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/b1.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/w2.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/b2.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/w3.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/b3.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/w4.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/taus/weights/b4.h"
+
 
 typedef ap_ufixed<16, 14> pt_t;
 typedef ap_fixed<10, 4> etaphi_t;
@@ -20,8 +34,6 @@ namespace L1TauEmu {
   //This way, the least significant bit of etaphi_t is exactly 0.01
   //Even though 0.01 is not a power of 2
   static float etaphi_base = 100. / 64;
-  static float z0_base = 0.05;
-  static float dxy_base = 0.05;
   typedef ap_ufixed<16, 14> pt_t;        // 1 unit = 0.25 GeV;
   typedef ap_fixed<10, 4> etaphi_t;      // 1 unit = 0.01;
   typedef ap_fixed<12, 6> detaphi_t;     // type for the difference between etas or phis
