+++ L1Trigger/Phase2L1ParticleFlow/src/newfirmware/puppi/linpuppi_ref.cpp
@@ -8,11 +8,11 @@
 #include "firmware/linpuppi_bits.h"
 #endif
 
-#include "../common/bitonic_hybrid_sort_ref.h"
-#include "../common/bitonic_sort_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/common/bitonic_hybrid_sort_ref.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/common/bitonic_sort_ref.h"
 
 #ifdef CMSSW_GIT_HASH
-#include "L1Trigger/Phase2L1ParticleFlow/src/dbgPrintf.h"
+#include "L1Trigger/Phase2L1ParticleFlow/interface/dbgPrintf.h"
 #else
 #include "../utils/dbgPrintf.h"
 #endif
